change to int type

+++ wk1/valid-c-code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(void) {
-    long long pointer = 1337;
+    int pointer = 1337;
     // Tell C that it's actually a pointer (to the address 1337) using type casting
     int *trust_me = (int*) pointer;
     // Dereference it and pray
